Fixed issue about the following: 1. .idea folder 2. min sdk version 3. repositories 4. removed picasso from build.gradle 5. added lateint for some variables 6. removed requirement for reading and writing to external storage, but not camera 7. changed most function to private function
