+++ editor/app/src/main/java/ca/ubc/eml/soiltopargraphy/editor/ui/infopanel/DescriptionPanelViewModel.kt
@@ -1,0 +1,9 @@
+package ca.ubc.eml.soiltopargraphy.editor.ui.infopanel
+
+import android.arch.lifecycle.ViewModel
+
+class DescriptionPanelViewModel : ViewModel() {
+    lateinit var name: String
+    lateinit var description: String
+    // TODO: Implement the ViewModel
+}
