+++ editor/app/src/main/java/ca/ubc/eml/soiltopargraphy/editor/ui/infopanel/InfoPanelFragment.kt
@@ -4,7 +4,6 @@ import android.arch.lifecycle.ViewModelProviders
 import android.net.Uri
 import android.os.Bundle
 import android.support.v4.app.Fragment
-import android.util.EventLogTags
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -13,7 +12,6 @@ import android.widget.TextView
 
 import ca.ubc.eml.soiltopargraphy.editor.R
 import com.bumptech.glide.Glide
-import org.w3c.dom.Text
 
 class InfoPanelFragment : Fragment() {
 
@@ -26,26 +24,10 @@ class InfoPanelFragment : Fragment() {
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
         val view  = inflater.inflate(R.layout.info_panel_fragment, container, false)
-        //find all four field
-        val descModel = activity?.run {
-            ViewModelProviders.of(this).get(DescriptionPanelViewModel::class.java)
-        }
-        val titleView = view.findViewById<TextView>(R.id.Info_panel_title)
-        val descriptionView = view.findViewById<TextView>(R.id.Info_panel_description)
-        val imageView = view.findViewById<ImageView>(R.id.ImageView)
-        val imageTitleView = view.findViewById<TextView>(R.id.Image_Title)
-        val imgmodel = activity?.run{
-            ViewModelProviders.of(this).get(ImagePanelViewModel::class.java)
-        }
-        loadUriInto(imgmodel?.uri,imageView)
-        titleView.text = descModel?.name
-        descriptionView.text = descModel?.description
-        imageTitleView.text = imgmodel?.imageTitle
-
         return view
     }
 
-    fun loadUriInto(uri: Uri?,view: ImageView) {
+    private fun loadUriInto(uri: Uri?,view: ImageView) {
         if(uri!=null) {
             Glide.with(context!!).load(uri).into(view)
         }
@@ -54,6 +36,24 @@ class InfoPanelFragment : Fragment() {
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
         viewModel = ViewModelProviders.of(this).get(InfoPanelViewModel::class.java)
+        val descModel = activity!!.run {
+            ViewModelProviders.of(this).get(DescriptionPanelViewModel::class.java)
+        }
+        val imgmodel = activity!!.run{
+            ViewModelProviders.of(this).get(ImagePanelViewModel::class.java)
+        }
+        viewModel.image = imgmodel.uri
+        viewModel.ImageTitle = imgmodel.imageTitle
+        viewModel.description = descModel.description
+        viewModel.name = descModel.name
+        val titleView = view?.findViewById<TextView>(R.id.Info_panel_title)
+        val descriptionView  = view?.findViewById<TextView>(R.id.Info_panel_description)
+        val imageView = view?.findViewById<ImageView>(R.id.ImageView)
+        val imageTitleView  = view?.findViewById<TextView>(R.id.Image_Title)
+        loadUriInto(viewModel.image, imageView!!)
+        titleView?.text = viewModel.name
+        descriptionView?.text = viewModel.description
+        imageTitleView?.text = viewModel.ImageTitle
         // TODO: Use the ViewModel
     }
