+++ editor/app/src/main/java/ca/ubc/eml/soiltopargraphy/editor/ui/infopanel/InfoPanelViewModel.kt
@@ -1,7 +1,12 @@
 package ca.ubc.eml.soiltopargraphy.editor.ui.infopanel
 
 import android.arch.lifecycle.ViewModel;
+import android.net.Uri
 
 class InfoPanelViewModel : ViewModel() {
+    lateinit var name: String
+    lateinit var description: String
+    lateinit var image: Uri
+    lateinit var ImageTitle: String
     // TODO: Implement the ViewModel
 }
